+++ packages/viewer/web/src/RichContentViewer.tsx
@@ -170,8 +170,6 @@ class RichContentViewer extends Component<
     contentState?: DraftContent
   ): ViewerContextType => {
     deprecateHelpers(helpers, config);
-    const version = Version.currentVersion;
-    const contentId = this.props.initialState?.ID;
     return {
       t,
       theme,
@@ -180,12 +178,7 @@ class RichContentViewer extends Component<
       relValue,
       customAnchorScroll,
       config,
-      helpers: {
-        ...helpers,
-        onViewerLoaded: args => helpers.onViewerLoaded?.({ ...args, version, contentId }),
-        onViewerAction: (pluginId, actionName, value) =>
-          helpers.onViewerAction?.(pluginId, actionName, value, contentId),
-      },
+      helpers,
       locale,
       disabled,
       seoMode,