+++ CHANGELOG.md
@@ -17,8 +17,41 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `ui-components`
+  - [#3273](https://github.com/wix/ricos/pull/3273) fix slider ui on rtl sites
+
+- `ricos-content`
+
+  - [#3276](https://github.com/wix/ricos/pull/3276) fromDraft conversion: handle invalid link types
+
+### ::house: Internal
+- `ricos-editor`
+  - [#3277](https://github.com/wix/ricos/pull/3277) add global context to ricos-modal
+
+
 </details>
 <hr/>
+
+## 8.65.16 (Dec 16, 2021)
+
+### :bug: Bug Fix
+
+- `ricos-content`
+
+  - [#3271](https://github.com/wix/ricos/pull/3271) fromDraft conversion: pass opts to nestedNodesConverters
+
+- `toolbars-new`
+
+  - [#3261](https://github.com/wix/ricos/pull/3261) add BI to link modal in new toolbar
+
+- `fullscreen`
+
+  - [#3274](https://github.com/wix/ricos/pull/3274) fullscreen opens on the correct image on mobile
+
+- `mentions`
+  - [#3235](https://github.com/wix/ricos/pull/3235) allow whitespaces
 
 ## 8.65.15 (Dec 16, 2021)